Add: Automatically apply the JaCoCo Gradle Plugin if required
