+++ plugin/src/main/kotlin/org/neotech/plugin/rootcoverage/utilities/AndroidGradlePluginExtensions.kt
@@ -1,7 +1,25 @@
 package org.neotech.plugin.rootcoverage.utilities
 
+import com.android.build.api.dsl.BuildType
+import com.android.build.api.variant.AndroidComponentsExtension
+import com.android.build.api.variant.Variant
 import org.gradle.api.Project
 import org.gradle.api.plugins.AppliedPlugin
+
+fun Project.onVariant(variantName: String, action: (variant: Variant?) -> Unit){
+    var variant: Variant? = null
+    val androidComponents = extensions.getByType(AndroidComponentsExtension::class.java)
+    androidComponents.onVariants {
+        if (it.name.replaceFirstChar(Char::titlecase) == variantName.replaceFirstChar(Char::titlecase)) {
+            variant = it
+        }
+    }
+    afterEvaluate {
+        action(variant)
+    }
+}
+
+
 
 fun Project.afterAndroidPluginApplied(notFoundAction: () -> Unit = {}, action: (AppliedPlugin) -> Unit) {
     var didExecuteBlock = false
