+++ plugin/src/test/test-fixtures/multi-module/app/build.gradle
@@ -1,6 +1,4 @@
 plugins {
-    id 'jacoco'
-
     // Cannot be applied using `alias(libs.plugins.androidApp)` since this test-fixture will run in
     // such a way that this plugin will already be available on the classpath
     id 'com.android.application'
