+++ plugin/src/test/test-fixtures/multi-module/build.gradle
@@ -1,6 +1,5 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 plugins {
-    id 'jacoco'
     id 'nl.neotech.plugin.rootcoverage'
 }
