+++ plugin/src/test/test-fixtures/multi-module/library_android/build.gradle
@@ -1,6 +1,4 @@
 plugins {
-    id 'jacoco'
-
     // Cannot be applied using `alias(libs.plugins.androidLibrary)` since this test-fixture will run in
     // such a way that this plugin will already be available on the classpath
     id 'com.android.library'
